pyproject.toml
[project]
name = "ttsbench"
version = "0.1.0"
description = "Local-first TTS benchmarking and training toolkit"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "Codex"}]
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.6.0",
  "pyyaml>=6.0.1",
  "soundfile>=0.12.1",
  "numpy>=1.26.0",
  "rich>=13.7.0",
  "sqlalchemy>=2.0.0",
  "tabulate>=0.9.0",
  "jiwer>=3.0.0",
]

[project.optional-dependencies]
metrics = [
  "faster-whisper>=1.0.0",
  "resemblyzer>=0.1.4",
]
train = [
  "torch>=2.2.0",
  "torchaudio>=2.2.0",
]
dev = [
  "pytest>=8.0.0",
]

[project.scripts]
ttsbench = "ttsbench.cli.main:app"

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
